% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setNodeProperties.R
\name{addContigProperty}
\alias{addContigProperty}
\title{addContigProperty}
\usage{
addContigProperty(contig, ko, property, all = FALSE, batchSize = 10000,
  add = FALSE)
}
\arguments{
\item{property}{list containing the property which you would want to insert, name of property in list is the name of the property in the}

\item{CONTIG}{the contigID}

\item{KO}{the KO the contig belongs to}
}
\description{
Adds property to contig node
}

