% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNodeProperties.R
\name{addKOProperty}
\alias{addKOProperty}
\title{addKOProperty}
\usage{
addKOProperty(ko, property)
}
\arguments{
\item{ko}{the koID}

\item{property}{the property which you want to add}
}
\description{
Adds property to ko node
}
