\name{adjacentPairs}
\alias{adjacentPairs}
\title{Convert igraph to ggvis object}
\description{Finds adjacent pairs of KOs or KO2CPD}
\usage{adjacentPairs(g, ko2ko = FALSE, ...)}
\arguments{
  \item{g}{graph object,}
  \item{ko2ko}{ko to ko}
  \item{\dots}{
}
}



\author{uesu}




\examples{
print("Still in progress")
}
