% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio.R
\name{allTrios}
\alias{allTrios}
\title{Find all trios surrounding the KO of interest}
\usage{
allTrios(KOI = "ko:K00001", koi = "ko:K00001", toUnique = TRUE,
  withDetails = FALSE, local = TRUE, contracted = TRUE, ...)
}
\arguments{
\item{KOI}{the ko id (for remote)}

\item{koi}{the ko id (for local)}

\item{toUnique}{if to return only unique (only applicable for remote)}

\item{withDetails}{with details (only applicable for remote)}

\item{local}{for doing the search locally, faster}

\item{contracted}{to contract anot (only applicable for local)}

\item{...}{the other args for dbquery}
}
\description{
Searches graphDB for all cases of trios surrounding the KO of interest
}
