% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateContigs.taxonomy}
\alias{annotateContigs.taxonomy}
\title{Adds taxonomic annotations of the contigs}
\usage{
annotateContigs.taxonomy(csv)
}
\arguments{
\item{csv}{FULL path pointing to outputfile from mapBlat package's blast2lcaIn fxn}
}
\description{
After loading into the graphDB contigs information we annotate contigs with their taxonomy sequences
}

