% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrMapping.R
\name{blatting}
\alias{blatting}
\title{blatting}
\usage{
blatting(ko, type, newblerInput)
}
\arguments{
\item{ko}{the ko of interest}

\item{type}{the}

\item{newblerInput}{path to newbler dir eg. /root in /root/K0000X/input/K0000X.1.fq}

\item{blatoutput}{path to store blat output in tabular blast format}
}
\description{
blatting
}
\examples{
\dontrun{

}
}
