% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTaxonomicTree.R
\name{buildE}
\alias{buildE}
\title{buildE sorts the tree into a data.frame}
\usage{
buildE(path, ma)
}
\arguments{
\item{path}{vector of integers representing taxid to common ancestor}

\item{ma}{matrix rbind with}
}
\description{
buildE sorts the tree into a data.frame
}
\keyword{internal}
