% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{categorize}
\alias{categorize}
\title{categorize}
\usage{
categorize(rs, ko)
}
\arguments{
\item{rs}{output from readStatusReader}
}
\description{
Adds 2 more columns
1. if the contig is chimeric or pure based on the read origins
2. if the readOrigin aligns with annotations
}
