% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contigInfo.R
\name{contigInfo}
\alias{contigInfo}
\title{contigInfo}
\usage{
contigInfo(contig, ko, withAnnotation = F)
}
\arguments{
\item{contig}{the contig}

\item{ko}{the ko bin which the contig was assembled from}
}
\description{
contigInfo function greps the contig related info such as whether its in the MDR
}

