% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{contigsSurvive.repeats}
\alias{contigsSurvive.repeats}
\title{contigsSurvive.repeats}
\usage{
contigsSurvive.repeats(readDepth, rs)
}
\arguments{
\item{readDepth}{the minimum depth}

\item{rs}{data.frame read.status, contig, readOrigin (taxid)}
}
\description{
contigsSurvive.repeats
}

