% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{contigsSurvive.repeats.readNum}
\alias{contigsSurvive.repeats.readNum}
\title{contigsSurvive.repeats.readNum}
\usage{
contigsSurvive.repeats.readNum(rs, thresholds = seq(1, 50, 1))
}
\arguments{
\item{rs}{data.frame read.status, contig, readOrigin (taxid)}

\item{thresholds}{the threshold window}
}
\description{
contigsSurvive.repeats.readNum
}

