% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{contigsSurvive.repeats.rpk}
\alias{contigsSurvive.repeats.rpk}
\title{contigsSurvive.repeats.rpk}
\usage{
contigsSurvive.repeats.rpk(rs, rpk = seq(0, 2, 0.01), koi = "K00927")
}
\arguments{
\item{rs}{data.frame read.status, contig, readOrigin (taxid)}

\item{rpk}{reads per kilo}

\item{koi}{the KO of interest}
}
\description{
contigsSurvive.repeats.rpk
}
