% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractMetab.R
\name{contractMetab}
\alias{contractMetab}
\title{contractMetab shrinks a metabolic network's KOs into non-redundant units}
\usage{
contractMetab(g)
}
\arguments{
\item{g}{metabolic graph, igraph object. vertices must have the Symbol property}
}
\value{
new non-redundant (gene) metabolic graph
}
\description{
Constracts KOs which share the same cpd product and substrates.
}

