% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpdname.R
\name{cpdname}
\alias{cpdname}
\title{Finds the details of the CPD when given its ID}
\usage{
cpdname(cpd = "C00022", ...)
}
\arguments{
\item{cpd}{the compound ID with or without the prefix is accepted either as a single or vector accepts with or without prefix}

\item{...}{allows for additional arguments use for dbquery}
}
\description{
Finds the details of the CPD when given its ID
}

