% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaMaps.R
\name{drawGradient}
\alias{drawGradient}
\title{drawGradient gives the color gradient}
\usage{
drawGradient(values, colors)
}
\arguments{
\item{values}{the score}

\item{colors}{the color}
}
\description{
sets the middle color at the median
}
\keyword{internal}
