% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{dynPlots}
\alias{dynPlots}
\title{dynPlot Diagnostic plots}
\usage{
dynPlots(dynList)
}
\arguments{
\item{dyn}{output from dynamicThreshold}
}
\description{
dynPlot Diagnostic plots
}
\examples{
\dontrun{
newblerDir = "~/simulation_fr_the_beginning/reAssemble/everybodyelse/data/newbler/"
c("ko:K00927","ko:K02316","ko:K02520") \%>\%
    gsub("ko:", "", .) \%>\% #folders dont begin with ko:
    lapply(dynamicThreshold(root=newblerDir)) \%>\% 
    dynPlot(dynList)
}
}

