% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio.local.R
\name{extractFromPath}
\alias{extractFromPath}
\title{extractFromPath}
\usage{
extractFromPath(edges, type = "cpd")
}
\arguments{
\item{edges}{the edgelist}

\item{type}{the type of nodes to extract}
}
\description{
given a set of edgelists extract out the CPD / KO node ids
}
\keyword{internal}
