% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHomology.R
\name{findHomology}
\alias{findHomology}
\title{searches for KO's homology assignments}
\usage{
findHomology(koi)
}
\arguments{
\item{koi}{the ko of interest}
}
\description{
`findHomology` seeks contigs which have been assigned via to 
the given KO of interest
used after homology search data has been uploaded into the db
}
