% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trioAnalysis.R
\name{findK}
\alias{findK}
\title{'findK to find the optimum number of Ks}
\usage{
findK(theMatrix, kmax = 10, ko)
}
\arguments{
\item{theMatrix}{two column data.frame with KS values of the before and after KOs in the reactions}

\item{kmax}{the max number of Ks to test for; defaults to 10}

\item{ko}{the ko of interest}
}
\value{
optimum number of Ks to choose
}
\description{
using cluster::clusGap
}

