% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacentPairs.R
\name{findNextKO}
\alias{findNextKO}
\title{FindNextKO}
\usage{
findNextKO(cpd, graph, originalKO, direction)
}
\arguments{
\item{cpd}{the intermediary cpd}

\item{graph}{metabolic graph}

\item{originalKO}{KO of interest}

\item{direction}{the direction in which to find}
}
\description{
Given a specied KO, one is suppose to find the next adjacent KO.
}

