% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTaxid.R
\name{findPerl}
\alias{findPerl}
\title{findPerl finds the path to the executable}
\usage{
findPerl(perl, verbose = "FALSE")
}
\arguments{
\item{perl}{full path to perl executable}

\item{verbose}{print extra}
}
\description{
Code taken from R package gdata
}
\keyword{internal}
