% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi2genus.R, R/sqlTaxid.R
\name{findPython}
\alias{findPython}
\title{findpython finds the path to the executable}
\usage{
findPython(perl, verbose = "FALSE")

findPython(perl, verbose = "FALSE")
}
\arguments{
\item{verbose}{print extra information}

\item{python}{full path to python executable}

\item{python}{full path to perl executable}

\item{verbose}{print extra}
}
\description{
Code taken from R package gdata

Code taken from R package gdata
}

