% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trioAnalysis.R
\name{findTrios}
\alias{findTrios}
\title{findTrios searches valid three-KO reactions resticting by the center KO and reports reaction clusters}
\usage{
findTrios(KOI, ks, toPrint = TRUE, outputFile, plotDir)
}
\arguments{
\item{KOI}{KOs-of-interest ie KOs above a selected amount of expression ie. highly expressed KOs and with D-statistics above that of the desired threshold}

\item{ks}{Precalculated data.frame output from ksCal fxn containing all KO's gene distribution  KS statistic when compared with whole sample's empirical gene distribution}

\item{toPrint}{conditional to print the results of the classification plots}

\item{outputFile}{the txt file to write the results of findTrios to}

\item{plotDir}{the location to save the diagnostics plots to}
}
\value{
data.frame of all reactions, corresponding clusters and the KS statistics for each KO and the selected Cluster
}
\description{
findTrios uses K-means clustering to identify reaction groups/clusters and the GAP statistic by Ryan Tibshirani to identity the best k implemented in the library cluster
Clustering algorithm uses the input matrix cmoposed of the the KS statistics of the KOs flanking the KO of interest.
}
\details{
The KS statistics is calculated based on the comparison of each KO's gene/contig expression distribution against the 'null' ie. 
empirical distribution against all genes in all contigs

In addition it also identifies reactions where flanking KOs have high gene diversity, given by low KS (<= 0.5)
}

