% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findtype}
\alias{findtype}
\title{findType finds KOs/compounds ID in metabolic graph}
\usage{
findtype(graph, type = c("c", "k"))
}
\arguments{
\item{graph}{graph object}

\item{type}{either find Cpds or KOs}
}
\value{
vector which nodes are of the certain type
}
\description{
findType finds KOs/compounds ID in metabolic graph
}

