% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio.local.R
\name{findV}
\alias{findV}
\title{findV finds the vertixID in the graph given its name}
\usage{
findV(name, g)
}
\arguments{
\item{name}{name of the vertex}

\item{g}{igraph object}
}
\description{
given the original graph find the vertiex ID of given name
}
