% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trioAnalysis.R
\name{getContigs}
\alias{getContigs}
\title{returns gene/contig infomation for all KOs}
\usage{
getContigs()
}
\value{
returns data.frame with columns: contigName, Freq_cDNA, rpkm_cDNA, Freq_gDNA, rpkm_gDNA, ko
}
\description{
queries graphDB and extracts information for all genes in all KOs, outputs this into a data.frame
}

