% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi2genus.R
\name{gi2rank}
\alias{gi2rank}
\title{gi2rank}
\usage{
gi2rank(blastTab, koi, sqlite3 = "sql3db")
}
\arguments{
\item{blastTab}{the tabular output file from DIAMOND}

\item{koi}{the KOs of interested eg K00001}

\item{sqlite3}{the sqlite3 db}
}
\description{
Parse tabular blast format of contigs from DIAMOND for their respective genusIDs
}
