% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{grep.cDNA}
\alias{grep.cDNA}
\title{grep.cDNA}
\usage{
grep.cDNA(reads)
}
\arguments{
\item{reads}{Shortreads object from bioconductor}
}
\description{
takes the cDNA from the input file (repeat of grepReads)
}
\keyword{internal}
