% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepgraph_cpd.R
\name{grepgraph.cpd}
\alias{grepgraph.cpd}
\title{Returns the metabolic graph given vector of KOs}
\usage{
grepgraph.cpd(cpds, fullGraph = FALSE, ...)
}
\arguments{
\item{fullGraph}{to output the full metabolic graph (yet to be implemented)}

\item{...}{additional dbquery parameters}

\item{cpd}{vector of cpds}
}
\description{
Takes in a vector list of cpd ids and generates a subset of metabolic graph as a igraph object
}
