% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphgraph_cpd.R
\name{grepgraph_cpd}
\alias{grepgraph_cpd}
\title{Generates metabolic network as IGRAPH object from input vector KOs
Returns the metabolic graph given vector of KOs}
\usage{
grepgraph_cpd(cpds, fullGraph = FALSE, ...)
}
\arguments{
\item{cpds}{vector of kos}

\item{fullGraph}{to export the whole of metabolism}

\item{...}{additional parameters for dbquery}
}
\description{
Takes in a vector list of KO ids and generates a metabolic graph as a igraph object
}

