% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrMapping.R
\name{grepReads}
\alias{grepReads}
\title{grepReads greps for cDNA or gDNA reads cause they are both grouped togehter in the assembly input}
\usage{
grepReads(reads, type = "cDNA")
}
\arguments{
\item{reads}{the ShortRead class object}

\item{type}{the pattern to look for}
}
\description{
grepReads greps for cDNA or gDNA reads cause they are both grouped togehter in the assembly input
}
\keyword{internal}
