% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig2ggplot.R
\name{ig2ggplot}
\alias{ig2ggplot}
\title{Convert igraph to ggplot2 object}
\usage{
ig2ggplot(g, dfOnly = TRUE, labels = FALSE, metab = TRUE, ...)
}
\arguments{
\item{g}{igraph object}

\item{dfOnly}{outputs the data.frame and not draw using ggvis}

\item{labels}{to show labels or not}

\item{metab}{it is a metabolic graph}

\item{...}{additional dbquery parameters}
}
\description{
Converts a metabolic igraph obj into a ggplot plot
}
