\name{ig2ggvis}
\alias{ig2ggvis}
\title{Convert igraph to ggvis object}
\description{Converts a igraph obj into a ggvis plot}
\usage{ig2ggvis(g, dfOnly = TRUE, ...)}
\arguments{
  \item{g}{graph object}
  \item{dfOnly}{df only}
  \item{\dots}{
}
}



\author{uesu}




\examples{
print("Still in Progress")
}
