% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph2gexf.R
\name{igraph2gexf}
\alias{igraph2gexf}
\title{Converts igraph obj two gexf
Function for converting igraph 2 gexf}
\usage{
igraph2gexf(mbgraph)
}
\arguments{
\item{mbgraph}{metabolic graph}
}
\description{
Converts igraph obj two gexf
Function for converting igraph 2 gexf
}

