% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCypher.R
\name{index}
\alias{index}
\title{Indexes the database for faster retrieval}
\usage{
index(label, property = "contig", cacheEnv)
}
\arguments{
\item{label}{label of the nodes you want to index}

\item{property}{the property key on which you would like to index on}

\item{cacheEnv}{stores details}
}
\description{
`index` indexes nodes based on the given property. Lets you do fast search 
for a node using a property field.
}
\examples{
\dontrun{
f = future({
    dbquery(query="CREATE INDEX ON :contigs(contig)", justPost = TRUE)
    dbquery(query="CREATE INDEX ON :contigs(bin)", justPost = TRUE)
    dbquery(query="CREATE INDEX ON :Taxon(taxid)", justPost = TRUE)
    dbquery(query="CREATE INDEX ON :cpd(cpd)", justPost = TRUE)
    message
}) \%plan\% multiprocess
}
}
