% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2ko.R
\name{ko2path}
\alias{ko2path}
\title{ko2path Finds all pathways related to the KO
Finds all associated pathways, and returns a data.frame with ko and pathway details}
\usage{
ko2path(ko = "K00001", ...)
}
\arguments{
\item{ko}{the ko ID int the form eg. K00001}
}
\description{
ko2path Finds all pathways related to the KO
Finds all associated pathways, and returns a data.frame with ko and pathway details
}

