% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trioAnalysis.R
\name{ksCal}
\alias{ksCal}
\title{ksCal generates KS statistics for aKO given the base distribution}
\usage{
ksCal(contigDF, baseDistribution, cores)
}
\arguments{
\item{contigDF}{data.frame with count frequency and rpkm of ALL KOs' contigs for both gDNA and cDNA obtained from query}

\item{baseDistribution}{the}

\item{cores}{number of cores to use}
}
\value{
data.frame with columns: ko, p.value and KS statistic (D)
}
\description{
runs a "two.sided" KS test and reports the KS and p value
}

