% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca}
\alias{lca}
\title{Finds the lowest common ancestor}
\usage{
lca(taxon1 = "10090", taxon2 = "9096", recurse = TRUE, ...)
}
\arguments{
\item{taxon1}{First taxon NCBI tax id}

\item{taxon2}{Second taxon NCBI tax id}

\item{recurse}{to keep moving up to find the LCA}

\item{...}{additional dbquery parameters}
}
\description{
Finds the lowest common ancestor
}

