% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listquery.R
\name{listquery}
\alias{listquery}
\title{Function for querying metamaps DB}
\usage{
listquery(query, params = FALSE, cypherurl = cacheEnv$cypher,
  user = cacheEnv$user, password = cacheEnv$password, ...)
}
\arguments{
\item{query}{cypher query}

\item{params}{string containing cypher query}

\item{cypherurl}{address of database, inclusive of ports}

\item{user}{database username}

\item{password}{database password}

\item{...}{accepts additional parameters from external function calls}
}
\description{
Takes variable queries with table output ie. no nodes no rels (simpler version of DBquery gives a list)
}

