% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{lookupTable}
\alias{lookupTable}
\title{generates `lookupTable` for filtering raw data post gene centric assembly}
\usage{
lookupTable(genesOfInterest = scg,
  annotations = "~/simulation_fr_the_beginning/out/template.csv", rs)
}
\arguments{
\item{genesOfInterest}{the kos of interest defaults to scg must be in the ko:KXXXXX format}

\item{annotations}{table containing the number of KOs for each taxID}

\item{rs}{processed readStatuses from newbler output}
}
\description{
returns three tables, 1. gene centric assembly readOrigin identities, 2. homologySearch assignment, 3 superimposed
}
\examples{
\dontrun{
rs = lapply(dynList, function(x) x$rs \%>\% mutate(ko = x$ko)) \%>\% do.call(rbind,.)
output = lookupTable(genesOfInterest = scg, annotations = "~/simulation_fr_the_beginning/out/template.csv", rs)
e
}
}

