% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{map}
\alias{map}
\title{map takes cDNA reads (fastQ format) and maps them onto contigs}
\usage{
map(reads, contigs, blatOutputFile)
}
\arguments{
\item{reads}{path to short reads file in fastq format}

\item{contigs}{path to contig file in fa format}
}
\description{
uses blat to map the mRNA to the contigs fasta and returns the tabular blast format
blat was chosen for its ability to show multiple hits
https://www.biostars.org/p/17613/
}
\examples{
\dontrun{
3
}
}

