% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrMapping.R
\name{mapContig}
\alias{mapContig}
\title{finds the location of the MDR on the contig}
\usage{
mapContig(contigid, seq, s, e)
}
\arguments{
\item{contigid}{the id of the contig}

\item{seq}{character string of the contig from the MSA}

\item{s}{starting loc}

\item{e}{ending loc}
}
\description{
finds the location of the MDR on the contig
}
