% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{mappingInfo}
\alias{mappingInfo}
\title{mappingInfo Maps contigs back onto origin gene (SIMULATION only)
And}
\usage{
mappingInfo(blasdf, root, koi, outputDir, query, rangesFile, genomesFile,
  blastFile, doBlast = F, ...)
}
\description{
mappingInfo Maps contigs back onto origin gene (SIMULATION only)
And
}

