% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrMapping.R
\name{mdrRanges}
\alias{mdrRanges}
\title{Returns contig ranges for those captured within the MDR}
\usage{
mdrRanges(ko, passDir)
}
\arguments{
\item{ko}{the KO of interest}

\item{passDir}{path to pAss outputs}
}
\description{
Header for contigs
contig00026 ref|YP_002288046.1| (ntRev) ## spanning:123 msaStart:653 msaEND:944 max10BPwindow:205.4 contigStart: 221 contigEnd: 421
Give GRanges output with seqname ko|contig00001
}
\examples{
\dontrun{
locsMDR = mdrRanges('K00927')
}
}
