\name{MetamapsDB-package}
\alias{MetamapsDB-package}
\alias{MetamapsDB}
\docType{package}
\title{MetamapsDB}
}
\description{For integration with neo4j's REST API}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetamapsDB}
\packageIndices{MetamapsDB}
}
\author{uesu}

Maintainer: \packageMaintainer{MetamapsDB}
}

\keyword{ package }


