% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nitrogenMetab.R
\name{nitrogenMetab}
\alias{nitrogenMetab}
\title{Igraph object for nitrogen metabolism}
\format{igraph object}
\source{
KEGG {ftp://bioinformatics.co.jp}
}
\description{
Nitrogen metabolism inclusive of the compounds and KOs
}

