% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2kingdom.R
\name{path2kingdom}
\alias{path2kingdom}
\title{List all intermediaries between taxa and the superkingdom it belongs to}
\usage{
path2kingdom(taxID = "79255", ...)
}
\arguments{
\item{taxID}{NCBI taxonomic id}

\item{...}{additional dbquery paramaters}
}
\description{
Finds the superkingdom given the taxid
}
\examples{
\dontrun{
path2kingdom(taxidID='79255')
Sending query

Query Completed

              name   taxid         rank
1         Gordonia   79255        genus
2         Theaceae   27065       family
3         Ericales   41945        order
4         asterids   71274     subclass
5     Pentapetalae 1437201      no rank
6       Gunneridae   91827      no rank
7   eudicotyledons   71240      no rank
8  Mesangiospermae 1437183      no rank
9    Magnoliophyta    3398      no rank
10   Spermatophyta   58024      no rank
11   Euphyllophyta   78536      no rank
12    Tracheophyta   58023      no rank
13     Embryophyta    3193      no rank
14  Streptophytina  131221      no rank
15    Streptophyta   35493       phylum
16   Viridiplantae   33090      kingdom
17       Eukaryota    2759 superkingdom

}
}
