% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2ko.R
\name{path2ko}
\alias{path2ko}
\title{Finds all KOs in a given pathway
Finds all KOs belonging to a Pathways}
\usage{
path2ko(pathway = "path:ko00010", ...)
}
\arguments{
\item{pathway}{The pathway ID}

\item{...}{additional paramters for dbquery}
}
\description{
Finds all KOs in a given pathway
Finds all KOs belonging to a Pathways
}
