% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%<>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs

lhs \%\$\% rhs
}
\description{
Pipe operator

Double Pipe operator

with Pipe operator
}
\keyword{internal}

