% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trioAnalysis.R
\name{plotClassification}
\alias{plotClassification}
\title{Plot clustering}
\usage{
plotClassification(matrix, ko)
}
\arguments{
\item{matrix}{input matrix for plotting needs before and after column}

\item{ko}{the koID for printing the KO name and ID to the diagnostic plot}
}
\description{
Plot clustering
}

