% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettifyGraph.R
\name{prettifyGraph}
\alias{prettifyGraph}
\title{returns metabolic graph with all the ornaments set
Adds details into the igraph object}
\usage{
prettifyGraph(g, vsize = c(1, 2), vcolor = c("grey", "red"),
  withText = TRUE, vtext = c(0.5, 1), layout = FALSE)
}
\arguments{
\item{g}{igraph Object}

\item{vsize}{the vertex size}

\item{vcolor}{length 2 vector for KO and compounds}

\item{withText}{Conditional to include labels in}

\item{vtext}{two vector text valuevsize}

\item{layout}{Whether to calculate the layoutvcolor}
}
\description{
returns metabolic graph with all the ornaments set
Adds details into the igraph object
}
