% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{rolling}
\alias{rolling}
\title{rolling A dynamic threshold for the killing redundant contigs}
\usage{
rolling(keptDF, wind = 15, interval = 10)
}
\arguments{
\item{keptDF}{dataframe for the #repeats left after thresholding, columnnames thresholds, repeats}

\item{wind}{window size}

\item{interval}{interval size}
}
\description{
outputs a data.frame with 2 columns 1.) rollingMean 2.) meanDepth
}

