% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{scg}
\alias{scg}
\title{singleCopyGenes'}
\format{31 KEGG ids ko:KXXXXX}
\usage{
scg(X, ev, nt, groups = NULL, mtype = c("symmetric", "laplacian",
  "stochastic"), algo = c("optimum", "interv_km", "interv", "exact_scg"),
  norm = c("row", "col"), direction = c("default", "left", "right"),
  evec = NULL, p = NULL, use.arpack = FALSE, maxiter = 300,
  sparse = igraph_opt("sparsematrices"), output = c("default", "matrix",
  "graph"), semproj = FALSE, epairs = FALSE, stat.prob = FALSE)
}
\description{
Vector list showing Single Copy Genes
}

