% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigmaMaps.R
\name{sigmaGraph}
\alias{sigmaGraph}
\title{Takes mbgraph and outputs as a shinyApp Obj
Users should encode the values into the igraph object
Works with older version of IGraph, maybe someone can choose to upgrade this further}
\usage{
sigmaGraph(igraphObj)
}
\arguments{
\item{igraphObj}{graph object after running prettifyGraph}
}
\description{
Takes mbgraph and outputs as a shinyApp Obj
Users should encode the values into the igraph object
Works with older version of IGraph, maybe someone can choose to upgrade this further
}

