% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{simulated}
\alias{simulated}
\title{finds all simulated genera for that particular KO}
\usage{
simulated(ko)
}
\arguments{
\item{ko}{koi}
}
\description{
`simulated` pull all simulated genomes
}
\keyword{internal}
