% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonname.R
\name{taxname}
\alias{taxname}
\title{Lists the taxonomic id's details}
\usage{
taxname(taxon = 5062, name = FALSE, ...)
}
\arguments{
\item{taxon}{The default is Aspergillus oryzae, accepts both name as well as}

\item{name}{search by name}

\item{...}{additional parameters for dbquery}
}
\description{
Lists the taxonomic id's details
}

