% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top500kos.R
\name{top500kos}
\alias{top500kos}
\title{The IDs of the top500 kos based on genomic fragment abundance}
\format{Chr vector 500 observations}
\description{
The IDs of the top500 kos based on genomic fragment abundance
}

