% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio.local.R
\name{trio.local}
\alias{trio.local}
\title{back2enz contracts and simplifies a metabolic subgraph (ie. neighborhood of a given}
\usage{
trio.local(koi, contracted = FALSE)
}
\arguments{
\item{koi}{the KO of interest}

\item{contracted}{conditional for contracting and simplifying graph ie. merging subunits}
}
\value{
data.frame containing all reactions
}
\description{
trio.local finds all trios surrounding a given KO
}
