% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNodeProperties.R
\name{addContigProperty}
\alias{addContigProperty}
\title{addContigProperty}
\usage{
addContigProperty(contig, ko, property, all = FALSE, batchSize = 10000,
  add = FALSE)
}
\arguments{
\item{contig}{the contigID}

\item{ko}{the KO the contig belongs to}

\item{property}{list containing the property which you would want to insert, name of property in list is the name of the property in the}

\item{all}{add all}

\item{batchSize}{the size of number of entries}

\item{add}{if to add property}
}
\description{
Adds property to contig node
}
