% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setNodeProperties.R
\name{addKOProperty}
\alias{addKOProperty}
\title{addKOProperty}
\usage{
addKOProperty(ko, property)
}
\arguments{
\item{ko}{the koID}

\item{property}{}
}
\description{
Adds property to ko node
}

