% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacentPairs.R
\name{adjacentPairs}
\alias{adjacentPairs}
\title{Finds adjacent pairs of KOs for actual biological reactions}
\usage{
adjacentPairs(g, ko2ko = FALSE)
}
\arguments{
\item{g}{graph object}

\item{ko2ko}{conditional to specify if one find the adjacent pairs directly from Ko 2 KO}
}
\description{
Finds adjacent pairs of KOs for actual biological reactions
}

