% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacentPairs.R
\name{adjacentPairs}
\alias{adjacentPairs}
\title{Finds adjacent pairs of KOs for actual biological reactions}
\usage{
adjacentPairs(g, ko2ko = FALSE, ...)
}
\arguments{
\item{g}{graph object}

\item{ko}{kos to find adjacent pairs for}
}
\description{
Finds adjacent pairs of KOs for actual biological reactions
}

