% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacentPairs.R
\name{adjacentPairs}
\alias{adjacentPairs}
\title{Given a igraph object, adjacentPairs finds
adjacent pairs of KOs}
\usage{
adjacentPairs(g, ko2ko = FALSE, ...)
}
\arguments{
\item{g}{graph object}

\item{ko2ko}{kos to find adjacent pairs for}
}
\description{
Given a igraph object, adjacentPairs finds
adjacent pairs of KOs
}

