% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawTaxonomicTree.R
\name{buildTree}
\alias{buildTree}
\title{buildTree - a taxonomic tree display function}
\usage{
buildTree(taxids)
}
\arguments{
\item{taxids}{vector of leave nodes}
}
\description{
draws the taxonomic tree of the given leave nodes
}

