% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTaxonomicTree.R
\name{buildTree}
\alias{buildTree}
\title{Taxonomic tree of the taxons}
\usage{
buildTree(taxids = c(287, 280), standardisedRanks = c("superkingdom",
  "phylum", "class", "order", "family", "genus", "species"))
}
\arguments{
\item{taxids}{vector of leave nodes}

\item{standardisedRanks}{the ranks to report. superkingdom phylum class order family genus species}
}
\description{
`buildTree` Returns taxonomic tree as a igraphObject of the taxids up to phylum
}
\examples{
\dontrun{
standardisedRanks  = c("superkingdom", "phylum", "class", "order","family", "genus", "species")
buildTree(taxids = c(287, 280), standardisedRanks )
}
}
