% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCypher.R
\name{connect}
\alias{connect}
\title{Settings for connecting to neo4j database
Function sets parameters such as the REST API URL, username and passwords
In the newer versions of neo4j default requires username and password}
\usage{
connect(url = "192.168.100.1", port = 7474, username = "neo4j",
  password = "neo4j")
}
\arguments{
\item{url}{url to graphDB}

\item{port}{port}

\item{username}{username}

\item{password}{password}
}
\description{
Settings for connecting to neo4j database
Function sets parameters such as the REST API URL, username and passwords
In the newer versions of neo4j default requires username and password
}

