% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contigInfo.R
\name{contigInfo}
\alias{contigInfo}
\title{contigInfo}
\usage{
contigInfo(ko, withAnnotation = FALSE, contig)
}
\arguments{
\item{ko}{the ko bin which the contig was assembled from}

\item{withAnnotation}{}

\item{contig}{contigID (optional)}
}
\description{
contigInfo function greps the contig related info such as whether its in the MDR
}

