\name{cpdname}
\alias{cpdname}
\title{List Pathways}
\description{Finds the details of the CPD when given its ID}
\usage{cpdname(cpd = "C00022", minimal = TRUE, ...)}
\arguments{
  \item{cpd}{the cpd ID; either as a single or vector accepts with or without prefix}
  \item{minimal}{The default is K00001, accepts both with or without the ko: prefix}
  \item{\dots}{
}
}



\author{uesu}




\examples{
#cpdname(cpd="C00022");
}
