% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbquery.R
\name{dbquery}
\alias{dbquery}
\title{Function for querying metamaps DB}
\usage{
dbquery(query, params = FALSE, cypherurl = cacheEnv$cypher,
  user = cacheEnv$user, password = cacheEnv$password, ...)
}
\arguments{
\item{query}{char string with cypher Query, rmbr to escape double quotes if any; read LUCENE}

\item{params}{a list object. ie params required by a cypher query, false if query requires no parameters; default}

\item{cypherurl}{address of database, inclusive of ports}

\item{user}{database username}

\item{password}{database password}

\item{...}{allows for additional arguments to be passed into dbquery}
}
\description{
Takes variable queries with table output ie. no nodes no rels
}

