% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{dynamicThreshold}
\alias{dynamicThreshold}
\title{dynamicThreshold tries to identify the lower bound converage value in order to remove low quality contigs.
Number of contigs tend to stablize at a value when removing reads below a certain readnum and we try to identify the min read required for that region of stability}
\usage{
dynamicThreshold(koi, mode = c("readnum", "rpk"), thresholding = c("rle",
  "rolling"), root)
}
\arguments{
\item{koi}{KO of interest}

\item{mode}{two options either readnum or coverage; coverage is experimental, use readnum}

\item{root}{folder to find}
}
\description{
dynamicThreshold tries to identify the lower bound converage value in order to remove low quality contigs.
Number of contigs tend to stablize at a value when removing reads below a certain readnum and we try to identify the min read required for that region of stability
}

