% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi2genus.R
\name{findPython}
\alias{findPython}
\title{findpython finds the path to the executable}
\usage{
findPython(python, verbose = "FALSE")
}
\arguments{
\item{python}{full path to python executable}

\item{verbose}{print extra information}
}
\description{
Code taken from R package gdata
}

