% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTaxid.R, R/utils.R
\name{findPython}
\alias{findPython}
\alias{findPython}
\title{findPython finds the path to the executable}
\usage{
findPython(python, verbose = "FALSE")

findPython(python, verbose = "FALSE")
}
\arguments{
\item{python}{full path to perl executable}

\item{verbose}{print extra}

\item{python}{full path to python executable}

\item{verbose}{print extra information}
}
\description{
Code taken from R package gdata

Code taken from R package gdata
}
