% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grepgraph.R
\name{grepgraph}
\alias{grepgraph}
\title{Returns the metabolic graph given vector of KOs}
\usage{
grepgraph(kos, fullGraph = FALSE, ...)
}
\arguments{
\item{kos}{vector of kos}

\item{fullGraph}{to output the full metabolic graph (yet to be implemented)}
}
\description{
Takes in a vector list of KO ids and generates a metabolic graph as a igraph object
}

