<<<<<<< HEAD
\name{ig2ggvis}
\alias{ig2ggvis}
\title{Convert igraph to ggvis object}

\usage{ig2ggvis(g, ...)}
\arguments{
  \item{g}{
}
  \item{\dots}{
}
}



\author{uesu}




\examples{
...
}
||||||| merged common ancestors
=======
\name{ig2ggvis}
\alias{ig2ggvis}
\title{Convert igraph to ggvis object}

\usage{ig2ggvis(g, dfOnly = TRUE, ...)}
\arguments{
  \item{g}{
}
  \item{dfOnly}{
}
  \item{\dots}{
}
}



\author{uesu}




\examples{
...
}
>>>>>>> refs/remotes/origin/master
