% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ig2ggvis.R
\name{ig2ggvis}
\alias{ig2ggvis}
\title{Convert igraph to ggvis object}
\usage{
ig2ggvis(g, dfOnly = TRUE, ...)
}
\arguments{
\item{g}{igraph object}

\item{dfOnly}{outputs the data.frame and not draw using ggvis}

\item{...}{additional dbquery parameters}
}
\description{
Converts a igraph obj into a ggvis plot
}

