\name{koname}
\alias{koname}
\title{List Pathways}
\description{Gives KO details when supplied with KO id}
\usage{koname(ko = "K00001", minimal = TRUE, ...)}
\arguments{
  \item{ko}{the koid}
  \item{minimal}{The default is K00001, accepts both with or without the ko: prefix}
  \item{\dots}{
}
}



\author{uesu}




\examples{
#pathwayDF = koname(ko="K00001");
}
