% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koname.R
\name{koname}
\alias{koname}
\title{Gives KO details when supplied with KO id}
\usage{
koname(ko = "K00001", minimal = TRUE, ...)
}
\arguments{
\item{ko}{the koid}

\item{minimal}{return minimal details}

\item{...}{additional dbquery parameters}
}
\description{
Gives KO details when supplied with KO id
}
