% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedata.frame.R
\name{make.data.frame}
\alias{make.data.frame}
\title{Convert columns with nested list structures into plain vectors}
\usage{
make.data.frame(df, string2factor = FALSE)
}
\arguments{
\item{df}{the data.frame output from dbquery}

\item{string2factor}{to store strings as factor}
}
\description{
Certain dbquerys will return data.frames with columns having lists, this function reduces this
}
\keyword{internal}
