% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weird.R
\name{mappingInfo}
\alias{mappingInfo}
\title{mappingInfo sequence analysis of contigs (for SIMULATION only)}
\usage{
mappingInfo(koi = "K00927", doBlast = FALSE, ...)
}
\arguments{
\item{koi}{the KO of interest}

\item{doBlast}{read from file or carry out blast}

\item{...}{root, outputDir, query, rangesFile, genomesFile, blastFile}
}
\description{
Used for mapping short reads against contigs using blast
}
\examples{
\dontrun{
blastfile = "~/simulation_fr_the_beginning/reAssemble/everybodyelse/out/simBlast/K00927/blast.tsv"
mappingInfo(blastfile, doBlast=FALSE)
}
}
