% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrMapping.R
\name{mapReads2MDR}
\alias{mapReads2MDR}
\title{overlaps}
\usage{
mapReads2MDR(ko, passDir, newblerInput)
}
\arguments{
\item{ko}{the ko of interest}

\item{passDir}{the directory of the pass project folder}

\item{newblerInput}{directory for newbler output files}
}
\description{
calculates overlaps
}
\examples{
\dontrun{
mapReads2MDR('K00927', passDir="./out", newblerInput="./newbler2")
}
}
