\name{MetamapsDB-package}
\alias{MetamapsDB-package}
\alias{MetamapsDB}
\docType{package}
\title{MetamapsDB}
\description{For integration with neo4j's REST API}
\details{
\tabular{ll}{Package: \tab MetamapsDB\cr
Maintainer: \tab uesu <wesley@bic.nus.edu.sg>\cr
Author: \tab uesu\cr
Version: \tab 0.01\cr
License: \tab GPL-2\cr
Title: \tab MetamapsDB\cr
Imports: \tab 
magrittr,
RJSONIO,
RCurl,
igraph,
rgexf,
magrittr,
ggvis\cr}
}
\author{uesu}

\keyword{ package }


