% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrogenMetab}
\alias{nitrogenMetab}
\title{nitrogenMetab}
\format{83 nodes and 251 edges. 50 KOs and 33 compounds}
\usage{
nitrogenMetab
}
\description{
Igraph object containing KOs and compounds involved in nitrogen metabolism
}
\keyword{datasets}

