% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2kingdom.R
\name{path2kingdom}
\alias{path2kingdom}
\title{List all intermediaries between taxa and the superkingdom it belongs to}
\usage{
path2kingdom(taxID = "79255", ...)
}
\arguments{
\item{taxID}{NCBI taxonomic id}

\item{...}{additional dbquery paramaters}
}
\description{
Finds the superkingdom given the taxid
}
