\name{path2ko}
\alias{path2ko}
\title{Finds all KOs belonging to a Pathways}
\description{Finds all KOs in a given pathway}
\usage{path2ko(pathway = "path:ko00010", ...)}
\arguments{
  \item{pathway}{The pathway ID}
  \item{\dots}{
}
}



\author{uesu}




\examples{
#kos.in.pathway = path2ko(pathway="path:K00982")
}
