% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prettifyGraph.R
\name{prettifyGraph}
\alias{prettifyGraph}
\title{returns metabolic graph with all the ornaments set
Adds details into the igraph object}
\usage{
prettifyGraph(g, vsize = c(1, 2), vcolor = c("grey", "red"),
  withText = TRUE, vtext = c(0.5, 1), layout = FALSE)
}
\arguments{
\item{g}{igraph Object}

\item{vsize}{the vertex size}

\item{vcolor}{length 2 vector for KO and compoundsprettifyGraph<- function}

\item{withText}{Conditional to include labels in plot(g,}

\item{vtext}{two vector text valuevsize    = c(1,2),}

\item{layout}{Whether to calculate the layoutvcolor   = c("grey", "red"),}
}
\description{
returns metabolic graph with all the ornaments set
Adds details into the igraph object
}

