% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{readStatusReader}
\alias{readStatusReader}
\title{readStatus stores the assignment details of a simulation into the graphDB
</rootDir/ko/454ReadStatus.txt> from newbler}
\usage{
readStatusReader(root, koi, mdr = FALSE)
}
\arguments{
\item{root}{the root directory}

\item{koi}{the ko directory}
}
\description{
readStatus stores the assignment details of a simulation into the graphDB
</rootDir/ko/454ReadStatus.txt> from newbler
}

