% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{readStatusReader}
\alias{readStatusReader}
\title{readStatusReader stores read assignment details from the simulation and assigns the contig a genome of origin}
\usage{
readStatusReader(root, koi, mdr = FALSE)
}
\arguments{
\item{root}{the root directory}

\item{koi}{the ko directory}
}
\description{
taken from /rootDir/ko/454ReadStatus.txt from newbler. 
We remove stray reads from other taxa if it only attributes to less than 10% of the contig, output
data.frame has column spanning set to 0 1 to show if its spanning or not
}
