% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{simulated}
\alias{simulated}
\title{finds all simulated genera for that particular KO}
\usage{
simulated(ko)
}
\arguments{
\item{ko}{koi}
}
\description{
`simulated` pull all simulated genomes
}
\examples{
/dontrun{

}
}
\keyword{#}
\keyword{%>%}
\keyword{/dontrun{}
\keyword{1}
\keyword{120961}
\keyword{13}
\keyword{2}
\keyword{3}
\keyword{33981}
\keyword{4}
\keyword{5}
\keyword{528}
\keyword{6}
\keyword{64}
\keyword{896}
\keyword{=}
\keyword{Freq}
\keyword{df}
\keyword{head}
\keyword{internal}
\keyword{ko}
\keyword{ko:K00001}
\keyword{simulated('ko:K00001)}
\keyword{taxid}
\keyword{}}
