% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlTaxid.R
\name{taxnam.sql}
\alias{taxnam.sql}
\title{taxname.sql}
\usage{
taxnam.sql(taxids, byID = TRUE, taxrank = "genus")
}
\arguments{
\item{taxids}{vector of integers/name representing the taxa of interest}

\item{byID}{search by ID}

\item{taxrank}{any of the common taxonomic ranks eg. genus}
}
\description{
faster way of querying the taxid names using structured SQL then querying the graphDB
}
