\name{taxname}
\alias{taxname}
\title{List Pathways}
\description{Lists the taxonomic id's details}
\usage{taxname(taxon = "5062", name = F, ...)}
\arguments{
  \item{taxon}{The default is Aspergillus oryzae, accepts both name as well as}
  \item{name}{
}
  \item{\dots}{
}
}



\author{uesu}




\examples{
#Asperigillus= taxname(taxon="5062");
}
