% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top500kos}
\alias{top500kos}
\title{Top500kos}
\format{A vector of 500 KOs based on KOs ranked by abundance, not sure if the vector is ranked.}
\usage{
top500kos
}
\description{
Dataset showing the top 500 KEGG Orthologs (KOS)
}
\keyword{datasets}

