\name{trio}
\alias{trio}
\title{#<< Function to find trios.}
\description{Finds trios}
\usage{trio(KOI = "ko:K00001", toFilter = FALSE, ...)}
\arguments{
  \item{KOI}{the KO of interest}
  \item{toFilter}{to filter or not}
  \item{\dots}{other arguments such as cypherurl}
}



\author{uesu}




\examples{
#...
}
