% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio.R
\name{trio}
\alias{trio}
\title{Function to find trios.}
\usage{
trio(KOI = "ko:K00001", toUnique = TRUE, withDetails = FALSE, ...)
}
\arguments{
\item{KOI}{the ko id}

\item{toUnique}{if to return only unique}

\item{withDetails}{with details}

\item{...}{the other args for dbquery}
}
\description{
Function to find trios.
}

